Fix force publish modal
If `data-toggle=modal` and `data-target` are used on a link element in
Bootstrap 3, then Bootstrap attempts to load the contents of the link's
href remotely, putting it into the modal container. There's no way of
disabling this. (The feature is being removed in Bootstrap 4)

We still want a link that works when js is broken or disabled. Use
`data-module=linked-modal` rather than `data-toggle=modal`

* The workaround doesn’t work in conjunction with the confirm dialogue,
remove the dialogue as the edition still requires a reason before it
can be published