Adding featured policies to org homepages
Organisations want to be able to pick the policies they feature on their
homepages. They also would like to be able to order the policies they
feature on their homepage. This is the first pass at adding featured
policies to organisations.

It pulls in the policies from the content store and saves references to
them so they can be displayed with their current name and slug on an org
homepage.